+++ lizard_blockbox/tests/test_models.py
@@ -73,6 +73,40 @@ class TestNamedReach(TestCase):
 
         self.assertEquals(namedreach.protection_levels, ["1250"])
 
+    def test_expanded_reaches(self):
+        """If there are two trajectories -- A, B, C, D and E, F, G, H --
+        and a river C, F, then it should be expanded to A, B, C, F, G, H."""
+        trajectory1 = factories.TrajectoryFactory.create(name=1)
+        trajectory2 = factories.TrajectoryFactory.create(name=2)
+
+        reaches = {
+            c: factories.ReachFactory.create(slug=c)
+            for c in "ABCDEFGH"}
+        for i, c in enumerate("ABCD"):
+            reaches[c].number = i
+            reaches[c].save()
+            trajectory1.reach.add(reaches[c])
+        for i, c in enumerate("EFGH"):
+            reaches[c].number = i
+            reaches[c].save()
+            trajectory2.reach.add(reaches[c])
+
+        river = factories.NamedReachFactory.create(name="test")
+
+        factories.SubsetReachFactory.create(
+            reach=reaches['C'], named_reach=river, km_from=100, km_to=200)
+        factories.SubsetReachFactory.create(
+            reach=reaches['F'], named_reach=river, km_from=201, km_to=300)
+
+        expanded_reaches = river.expanded_reaches()
+        self.assertEquals(len(expanded_reaches), 6)
+        self.assertEquals(expanded_reaches[0].slug, 'A')
+        self.assertEquals(expanded_reaches[1].slug, 'B')
+        self.assertEquals(expanded_reaches[2].slug, 'C')
+        self.assertEquals(expanded_reaches[3].slug, 'F')
+        self.assertEquals(expanded_reaches[4].slug, 'G')
+        self.assertEquals(expanded_reaches[5].slug, 'H')
+
 
 class TestVertex(TestCase):
     def test_years(self):